006-module/requirements.txt
pandas>=1.5.0
sentence-transformers>=2.2.0
qdrant-client>=1.7.0
duckduckgo-search>=3.9.0
requests>=2.28.0
python-dotenv>=0.19.0
pydantic>=2.0.0
tqdm>=4.65.0
nbformat
gradio>=4.44.1,<5.0.0
typer>=0.9.0,<0.13
click>=8.1.3,<9.0.0
rich>=13.0.0
openai>=1.0.0
fastapi>=0.111.0
starlette>=0.37.2
